Started implementation of driving and testing motors
I updated driving command to use the drivetrain as Arcade Drive with the
driver joystick set in the OI. I implemented the Drivetrain to allow for
driving as well as currently being able to test motor a by pressing
button 3 (set in OI) to have the motor run at .25. I have not put much
effort into trying to fix how motor controller situation being that I
believe they are still programmed for PWM and need to be changed to CAN. 

Testing motors B and C has not been set up yet but it will be identical
to the implementation to testing motor A. 

Currently a lot of data is being pushed to SmartDashboard, likely more
than needed but I decided I would rather do overkill than have to set up
more when I used RobotBuilder.