Solenoid programmed error free
With the code just finished I have yet to test if the code works but I
can say the code is completely finished for a solenoid (this is not
currently set up for a double solenoid), has a function in the
Pneumatics subsystem to test the solenoid that just toggles between
positions, and a button to run the command the new command
"UseSolenoid". Currently it is set up to test specifically the solenoid
set up but I am looking at making the function so we can test any
solenoid inputted.

It is on the idea of scalability. Why have several nearly identical
functions when you can have one to test the them all? Currently it would
still mean several commands made using only the one function, however,
it could also be reduced to a single command by choosing which solenoid
in the SmartDashboard just like we would like to choose which Autonomous
mode. This would also directly apply to testing motors, if we
implemented similar commands and functions we could test all our motors
through the SmartDashboard with very little code to do it all.