Implemented Solenoid Test Function and Command For Any Solenoid
I have added the function(uncommented actually): TestaSolenoid to
Pneumatics, that allows testing of any solenoid inputted. Additionally I
added and implemented the UseASolenoid command so that a single command
is used to test any solenoids. 

TODO: Implement the same things for motors so we only use a single
command and function to test all the motors.