Drivetrain fix, Implemented Live Window, and fixed Test a Solenoid on SD
Drivetrain's numbers were off a bit. One number should have been a
negative and it should have divided by sqrt(2) not multiplied. 

LiveWindow was implemented because of the update in the Wpilib allowing
CANTalons to be used as actuators. 

Test a Solenoid on the SD was broken or I was using the wrong function,
do not recall but now it is correct.